+++ libs/ghostos/ghostos/libraries/project/directory_impl.py
@@ -12,6 +12,8 @@ from ghostos_common.helpers import generate_directory_tree, yaml_pretty_dump, ge
 from ghostos_common.helpers.files import DescriptionsGetter
 from ghostos_moss import moss_runtime_ctx
 from pydantic import Field
+from contextlib import contextmanager
+import time
 
 
 class DirectoryData(YamlConfig):
@@ -28,10 +30,16 @@ class DirectoryData(YamlConfig):
         default=None,
         description="the editing filename relative to the current directory",
     )
+    updated: int = Field(
+        default=0,
+        description="the updated timestamp",
+    )
 
     def save_to(self, path: pathlib.Path) -> None:
         if not path.is_dir():
             raise NotADirectoryError(f'{path} is not a directory')
+        now = int(time.time())
+        self.updated = now
         content = yaml_pretty_dump(self.model_dump(exclude_defaults=True))
         file = path.joinpath(self.relative_path)
         with open(file, "w") as f:
@@ -131,9 +139,14 @@ class FileImpl(File):
 
         modulename = get_module_fullname_from_path(str(self.path), use_longest_match=True)
         if modulename is not None:
-            py_info = f"\n\nfile is python module `{modulename}`."
+            py_info = f"\nfile is python module `{modulename}`."
+            with moss_runtime_ctx(modulename) as rtm:
+                imported_prompt = rtm.prompter().get_imported_attrs_prompt()
+                if imported_prompt:
+                    py_info += "imported attr information are: \n```python\n" + imported_prompt + "\n```"
+
             return (f'content of file `{self.path}` are:'
-                    f'\n\n<content length="{length}">\n{content}{suffix}\n</content>{py_info}')
+                    f'\n\n<content length="{length}">\n{content}{suffix}\n</content>\n{py_info}')
 
         return f'<content length="{length}">\n{content}{suffix}\n</content>'
 
@@ -277,13 +290,16 @@ class DirectoryImpl(Directory):
         ".gitignore",
         DirectoryData.relative_path,
         "__pycache__",
+        '.git/',
+        '.idea/',
     ]
 
-    def __init__(self, path: pathlib.Path, ignores: List[str] = None, relative: Union[str, None] = None):
+    def __init__(self, path: pathlib.Path, ignores: List[str] = None):
+        editing_file = None
+        if not path.is_dir():
+            editing_file = path.name
+            path = path.parent
         self.path = path
-        if relative is None:
-            relative = self.path.absolute()
-        self.relative = relative
         self.data = DirectoryData.get_from(path)
         self.ctx = self.data.get_dev_context('.')
         if not self.ctx.desc:
@@ -301,6 +317,7 @@ class DirectoryImpl(Directory):
             if ignore.startswith('#'):
                 continue
             self._ignores.append(ignore)
+        self.focus(editing_file)
 
     def get_ignores(self) -> List[str]:
         return self._ignores
@@ -309,9 +326,10 @@ class DirectoryImpl(Directory):
         editing_context = "empty"
         if self.data.editing:
             try:
-                editing_context = self.edit(self.data.editing).read(line_number=True, detail=False)
-            except Exception:
-                pass
+                file = self.edit(self.data.editing)
+                editing_context = file.read(line_number=True, detail=True)
+            except Exception as e:
+                editing_context = f"can not read content on error {e}"
 
         return f"""
 full context of the Directory instance:
@@ -428,24 +446,31 @@ all the available dev contexts from name (or path) to description are:
     def save_dev_contexts(self):
         self.data.save_to(self.path)
 
+    def save_data(self) -> None:
+        self.data.save_to(self.path)
+
 
 if __name__ == "__main__":
     current = pathlib.Path(__file__)
     d = DirectoryImpl(current.parent)
-    d.focus(current.name)
-    print(d.lists())
-    print(d.ctx.full_context())
-    d.describe("abcd.py", "abstract classes")
-    d.save_dev_contexts()
+    with d:
+        d.focus(current.name)
+        print(d.lists())
+        print(d.ctx.full_context())
+        d.describe("abcd.py", "abstract classes")
+        d.save_dev_contexts()
 
-    print("+++++")
-    print(d.dev_contexts())
+        print("+++++")
+        print(d.dev_contexts())
 
-    print("+++++")
-    print(d.full_context())
+        print("+++++")
+        print(d.full_context())
 
-    editing = d.edit(current.name)
-    print("+++++++++++ path", editing.path)
+        editing = d.edit(current.name)
+        print("+++++++++++ path", editing.path)
 
-    editing.write("hello", append=True)
-    insert_content(editing.path, "world", -1, -1)
+        editing.write("hello", append=True)
+        insert_content(editing.path, "world", -1, -1)
+
+hello
+world