+++ libs/ghostos/ghostos/libraries/project/project_impl.py
@@ -1,5 +1,4 @@
 from typing import Optional
-from typing_extensions import Self
 
 from ghostos.abcd import Operator, Session
 from ghostos.libraries.project.abcd import ProjectManager
@@ -10,7 +9,7 @@ from ghostos.libraries.pyeditor.simple_module_editor import SimplePyModuleEditor
 from pydantic import BaseModel, Field
 from ghostos_common.helpers import generate_import_path
 from ghostos_common.prompter import PromptObjectModel, TextPOM
-from ghostos_moss import MossRuntime, Modules
+from ghostos_moss import Modules
 from ghostos_container import Container, Provider
 import pathlib
 
@@ -82,11 +81,10 @@ class ProjectManagerImpl(ProjectManager, PromptObjectModel):
         self.working: DirectoryImpl = DirectoryImpl(
             working_path,
             ignores=None,
-            relative=str(working_path.relative_to(root_dir)),
         )
 
     def work_on(self, dir_path: str) -> Operator:
-        dir_root = self.working.path
+        dir_root = self.root.path
         if dir_path == '.' or dir_path == '~':
             dir_root = self.root.path
             dir_path = dir_path[1:]